Rename highcharts module
